+++ src/pages/Login/LoginView.tsx
@@ -3,7 +3,7 @@ import {useRecoilValue, useSetRecoilState} from 'recoil';
 import {Container, Title} from './Login.styled';
 import {langState} from '@modules/atoms';
 import {Lang} from '@type/index';
-import {LoginString} from '@utils/constants';
+import {LoginString} from '@utils/constants/strings';
 
 export const LoginView = () => {
   const lang = useRecoilValue<Lang>(langState);
