+++ src/pages/TicketPage/TicketPageView.tsx
@@ -3,7 +3,7 @@ import {useRecoilValue} from 'recoil';
 
 import {langState} from '@modules/atoms';
 import {Container, Logo} from './styles';
-import {TicketPageString} from '@utils/constants';
+import {TicketPageString} from '@utils/constants/strings';
 
 const TicketPageView = () => {
   const lang = useRecoilValue(langState);
